backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai==0.8.3
yt-dlp>=2025.1.26
aiofiles==23.2.1
python-multipart==0.0.12
pydantic==2.9.2
httpx==0.27.2
openai
youtube-transcript-api>=0.6.3
